Incorrect keyboard
/*
Incorrect keyboard Suman's keyboard got damaged as he dropped it from a considerable distance. As a result instead of K the keyboard may sometimes display T (but will never display K instead of T)Similarly instead of G, the keyboard may sometimes display D, instead of R sometimes it may display L and sometimes instead of R it displays F.
Because of this Suman has to think of all possible words for a displayed word (by typing using his keyboard). 
The program to be written should accept a word W which is displayed as a result of typing on Suman's keyboard and should calculate how many possible words it could mean (if typed using a properly working keyboard) and type the count as the output.

Boundary Conditions:The length of the input word is between 2 and 100.

Input Format:
First line will contain the word W.

Output Format:
The count of possible words that the input word (typed using Suman's keyboard) could mean.

Sample Input/Output:
Example 1:
Input:
FILIPEK
Output: