Finding same set of digits in Java
 Given a number, find next greater number with the same set of digits as original number.
For example,
if input number is 1234, then output should be 1243
if input number is 1342, then output should be 1423
if input number is 534976, then output should be 536479

if input number is 4321, then output should be next greater number not possible. (As input number is the greatest number that can be formed with the digits.